requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
channels==4.3.1
channels-redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
celery==5.5.3
cryptography==46.0.1
django-celery-beat==2.8.1
django-celery-results==2.6.0
django-cors-headers==4.8.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
django-storages==1.14.6
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
jsonschema==4.25.1
MarkupSafe==3.0.2
numpy==2.3.3
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
Pillow==11.3.0
pyparsing==3.2.4
pydantic==2.11.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
sqlparse==0.5.3
stripe==12.5.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
urllib3==2.5.0
whitenoise==6.10.0
yarl==1.20.1
ccxt==4.5.4  # Cryptocurrency Trading Library4
ccxt[pro]
influxdb-client==1.49.0
scikit-learn==1.7.2 
tensorflow==2.20.0
ta-lib==0.6.7  # Technical Analysis Library
ta==0.11.0      # Alternative TA library
python-binance==1.0.29
websocket-client==1.6.0
asyncio-mqtt==0.16.2
flower
channels[daphne]
daphne
transformers
torch
textblob
accelerate
huggingface_hub[hf_xet]
tf-keras
spacy
coverage
bertopic
sentence-transformers
umap-learn
hdbscan
langdetect
llama-index
llama-index-core
llama-index-vector-stores-faiss
llama-index-embeddings-huggingface
llama-index-llms-openai
llama-index-llms-anthropic
llama-index-llms-ollama
llama-index-llms-groq
faiss-cpu
openai
anthropic
groq